# api_docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/tournaments/{tournamentId}/rounds/{roundId}/matches/{matchId}:
    get:
      tags:
      - tournament-rest-controller
      operationId: getRoundMatch
      parameters:
      - name: tournamentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: roundId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: matchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MatchUp_MatchDetails'
    put:
      tags:
      - tournament-rest-controller
      operationId: updateMatch
      parameters:
      - name: tournamentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: matchId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchUp_MatchDetails'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MatchUp_MatchDetails'
  /api/tournaments/{id}:
    get:
      tags:
      - tournament-rest-controller
      operationId: getTournament
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Tournament_TournamentDetails'
    put:
      tags:
      - tournament-rest-controller
      operationId: replaceTournament
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tournament_TournamentDetails'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Tournament_TournamentDetails'
...
  /api/tournaments/{tournamentId}/rounds:
    get:
      tags:
      - tournament-rest-controller
      operationId: getTournamentRounds
      parameters:
      - name: tournamentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rounds_RoundDetails'
    post:
      tags:
      - tournament-rest-controller
      operationId: createRounds
      parameters:
      - name: tournamentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Rounds_RoundDetails'
  /api/tournaments/{tournamentId}/participants:
    get:
      tags:
      - tournament-rest-controller
      operationId: getTournamentParticipants
      parameters:
      - name: tournamentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team_TournamentDetails'
    post:
      tags:
      - tournament-rest-controller
      operationId: addParticipantToTournament
      parameters:
      - name: tournamentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Team_TournamentDetails'
  /api/tournaments/:
    get:
      tags:
      - tournament-rest-controller
      operationId: getTournaments
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tournament_TournamentDetails'
    post:
      tags:
      - tournament-rest-controller
      operationId: createTournament
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tournament_TournamentDetails'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Tournament_TournamentDetails'
...
  /api/tournaments/{tournamentId}/rounds/{roundId}:
    get:
      tags:
      - tournament-rest-controller
      operationId: getTournamentRound
      parameters:
      - name: tournamentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: roundId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Rounds_RoundDetails'
  /api/tournaments/{tournamentId}/rounds/{roundId}/matches:
    get:
      tags:
      - tournament-rest-controller
      operationId: getRoundMatches
      parameters:
      - name: tournamentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: roundId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchUp_MatchDetails'
  /api/tournaments/{tournamentId}/participants/{teamId}:
    get:
      tags:
      - tournament-rest-controller
      operationId: getTournamentParticipant
      parameters:
      - name: tournamentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: teamId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Team_TournamentDetails'
  /api/tournaments/pages:
    get:
      tags:
      - tournament-rest-controller
      operationId: getTournametPage
      parameters:
      - name: numPage
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tournament_TournamentDetails'
...
  /api/teams/pages:
    get:
      tags:
      - team-rest-controller
      operationId: getTournametPage_1
      parameters:
      - name: numPage
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team_TeamDetails'
...
components:
  schemas:
...
    Game_MatchDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    MatchUp_MatchDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        team1:
          $ref: '#/components/schemas/Team_MatchDetails'
        team2:
          $ref: '#/components/schemas/Team_MatchDetails'
        score1:
          type: integer
          format: int32
        score2:
          type: integer
          format: int32
        isPlayed:
          type: boolean
        winner:
          $ref: '#/components/schemas/Team_MatchDetails'
        played:
          type: boolean
    Team_MatchDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        games:
          type: array
          items:
            $ref: '#/components/schemas/Game_MatchDetails'
        requests:
          type: array
          items:
            type: integer
            format: int64
    Game_TournamentDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Rounds_TournamentDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        numRound:
          type: integer
          format: int32
        evenRound:
          type: boolean
    Team_TournamentDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        games:
          type: array
          items:
            $ref: '#/components/schemas/Game_TournamentDetails'
        requests:
          type: array
          items:
            type: integer
            format: int64
    Tournament_TournamentDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        currentPlayers:
          type: integer
          format: int32
        maxPlayers:
          type: integer
          format: int32
        iniDate:
          type: string
        endDate:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Team_TournamentDetails'
        winner:
          $ref: '#/components/schemas/Team_TournamentDetails'
        rounds:
          type: array
          items:
            $ref: '#/components/schemas/Rounds_TournamentDetails'
        game:
          $ref: '#/components/schemas/Game_TournamentDetails'
        roundNumber:
          type: integer
          format: int32
        started:
          type: boolean
        finished:
          type: boolean
        admin:
          $ref: '#/components/schemas/User_TournamentDetails'
    User_TournamentDetails:
      type: object
      properties:
        encodedPassword:
          type: string
        roles:
          type: array
          items:
            type: string
    Game_TeamDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Team_TeamDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        tournaments:
          type: array
          items:
            $ref: '#/components/schemas/Tournament_TeamDetails'
        games:
          type: array
          items:
            $ref: '#/components/schemas/Game_TeamDetails'
        creator:
          $ref: '#/components/schemas/User_TeamDetails'
        users:
          type: array
          items:
            $ref: '#/components/schemas/User_TeamDetails'
        admins:
          type: array
          items:
            $ref: '#/components/schemas/User_TeamDetails'
        requests:
          type: array
          items:
            type: integer
            format: int64
    Tournament_TeamDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        maxPlayers:
          type: integer
          format: int32
        iniDate:
          type: string
        endDate:
          type: string
        game:
          $ref: '#/components/schemas/Game_TeamDetails'
        roundNumber:
          type: integer
          format: int32
        started:
          type: boolean
        finished:
          type: boolean
        admin:
          $ref: '#/components/schemas/User_TeamDetails'
    User_TeamDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        encodedPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        riot:
          type: string
        blizzard:
          type: string
        psn:
          type: string
        xbox:
          type: string
        steam:
          type: string
    Game_RoundDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    MatchUp_RoundDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        isPlayed:
          type: boolean
        winner:
          $ref: '#/components/schemas/Team_RoundDetails'
        played:
          type: boolean
    Rounds_RoundDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        matches:
          type: array
          items:
            $ref: '#/components/schemas/MatchUp_RoundDetails'
        numRound:
          type: integer
          format: int32
        evenRound:
          type: boolean
    Team_RoundDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        games:
          type: array
          items:
            $ref: '#/components/schemas/Game_RoundDetails'
        requests:
          type: array
          items:
            type: integer
            format: int64
...